+++ statcalc/variance.go
@@ -1,10 +1,10 @@
 package statcalc
 
-func CalculateVariance(data []int, mean int) int {
-	sumOfSquares := 0
+func CalculateVariance(data []float64, mean float64) float64 {
+	sumOfSquares := 0.0
 	for _, value := range data {
 		diff := value - mean
 		sumOfSquares += diff * diff
 	}
-	return sumOfSquares / len(data)
+	return sumOfSquares / float64(len(data))
 }
